Change schemas.Column to use int64
The largest size of a column in SQL is above the largest integer that
can be stored in an int/int32. Running go on a 32bit machine will result
in ints being mapped to int32 and thus interpretting the schema will
fail.

In this PR we change the schema.Column to use int64 which will allow
Gitea to fix https://github.com/go-gitea/gitea/issues/20161

Signed-off-by: Andrew Thornton <art27@cantab.net>